requirements.txt
httpx
beautifulsoup4
motor
pycryptodome
fake-useragent
